requirements.txt
alembic==1.11.1
altair==4.1.0
celery==5.2.7
cryptography==41.0.6
datasketch==1.5.0
dfir-unfurl==20230901
opensearch-py==2.3.1
Flask==2.3.2
flask_bcrypt==1.0.1
flask_login==0.6.2
flask_migrate==4.0.4
flask_restful==0.3.10
flask_sqlalchemy==3.0.3
flask_wtf==1.1.1
google-auth==1.7.0
google_auth_oauthlib==0.4.1
gunicorn==19.10.0  # Note: Last version to support py2
numpy==1.23.4
oauthlib==3.1.0
pandas==1.5.0
PyJWT==2.4.0
python_dateutil==2.8.1
PyYAML==6.0.1
redis==4.4.4
requests==2.31.0
sigmatools==0.19.1 ; python_version > '3.4'
six==1.12.0
SQLAlchemy==1.4.48
Werkzeug==2.3.4
WTForms==3.0.1
xlrd==1.2.0
tabulate==0.9.0
markdown==3.2.2
networkx==2.5
prometheus-client==0.16.0
prometheus-flask-exporter==0.22.4
decorator==5.0.5
geoip2==4.2.0
jsonschema==4.17.3
